+++ CHANGELOG.md
@@ -16,6 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [1.0.0] - 2024-03-29
+
+### Added
+
+- Sorting the archives by date or name
+- Editing the archives description and name
+
 ## [0.6.1] - 2024-03-29
 
 ### Fixed